Split up Metaculus
