+++ ergo/platforms/metaculus/__init__.py
@@ -1,0 +1,2 @@
+from .metaculus import Metaculus
+from .question import MetaculusQuestion
